Moves IPasswordService implementation to Orchard Users Creates Extensions to share management of Password operations across Services
