+++ src/Orchard/Security/IPasswordService.cs
@@ -1,6 +1,6 @@
 ﻿namespace Orchard.Security {
     public interface IPasswordService : IDependency {
         bool Equals(PasswordContext context, string plaintextPassword);
-        string ComputeHashBase64(string hashAlgorithmName, byte[] saltBytes, string password);
+        //string ComputeHashBase64(string hashAlgorithmName, byte[] saltBytes, string password);
     }
 }
